Correctly getting values from Chef::Config
Chef::Config extends MixLab::Config, it isn't an actual hash.
MixLab::Config provides a bracket syntax and a dot syntax for getting
values.  I would normally prefer the dot syntax (to make it clear that
we're not grabbing keys from a hash) but I used the bracket syntax here
for consistancy with the rest of the knife-server codebase.